Day 5.
Fleshing out the INTCODE computer a bit more. Got stuck for a bit when I
forgot that the part1 code was modifying storage that I was then passing
to part2, so the initial instruction (whatever you pass as input becomes
one of the first instructions run) was wrong, and it was using
jump-if-false instead of jump-if-true because an extra 1 (the input in
part1) was added to the instruction.