Day 7!
This was a fun one; part one was just a "do a bunch of permutations of
things", but part 2 was basically: run a bunch of machines in series.

In this case, I did it as a simple cooperative multitasking thing,
because the wait-for-input behavior is so simple. Now the machine state
is stored in struct, and the machine's input is unowned, so it can be
either externally passed in OR hooked up to another machine, as the case
is here. Then when a machine goes to read input and doesn't have enough,
it steps the pc back (to the read-input instruction) and then halts with
a kWaitingForInput. Then we just execute all the machines until they all
halt (via kHalt, not waiting) and print the last output of the last
machine.

This was one of those really rare times where I wrote all the Part 2
bits at once and, once it compiled, it worked the first time. Not too
bad :)